Improve `CodeEditor` wrapper (#114500)
# Conflicts:
#	packages/kbn-optimizer/limits.yml
