+++ packages/kbn-optimizer/limits.yml
@@ -108,7 +108,7 @@ pageLoadAssetSize:
   data: 491273
   dataViews: 42532
   fieldFormats: 65209
-  kibanaReact: 99422
+  kibanaReact: 84422
   uiActions: 35121
   dataEnhanced: 24980
   embeddable: 87309
