+++ src/plugins/kibana_react/public/code_editor/languages/constants.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export { LANG as CssLang } from './css/constants';
+export { LANG as MarkdownLang } from './markdown/constants';
+export { LANG as YamlLang } from './yaml/constants';
+export { LANG as HandlebarsLang } from './handlebars/constants';
